Address feedback on naming and constraints

+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/domain/TopicMessage.java
@@ -45,5 +45,5 @@ public class TopicMessage {
 
     private int topicNum;
 
-    private long runningHashVersion;
+    private int runningHashVersion;
 }
